requirements.txt
Flask==1.0.3
asn1crypto==1.0.0
bcrypt==3.1.6
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nexmo==2.4.0
Pillow==6.1.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.6
Werkzeug==0.15.4
WTForms==2.2.1
gunicorn==19.9.0


